+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - refactored and set up for DiGraph support
 - lighting improvements
 - debugged edge and node labels, colors
+- renamed main class from NxPlot to Nx3D
 
 # Sat Sep 17 2022
 22.9.5